+++ src/config.js
@@ -1,9 +1,8 @@
 export const firebaseConfig = {
-  apiKey: "AIzaSyDOTe8SVN_5WGtuKdXQoQy5slhAq3kEIbs",
-  authDomain: "findy-id.firebaseapp.com",
-  databaseURL: "https://findy-id.firebaseio.com",
-  projectId: "findy-id",
-  storageBucket: "findy-id.appspot.com",
-  messagingSenderId: "943881521886"
-};
-
+  apiKey: 'AIzaSyDOTe8SVN_5WGtuKdXQoQy5slhAq3kEIbs',
+  authDomain: 'findy-id.firebaseapp.com',
+  databaseURL: 'https://findy-id.firebaseio.com',
+  projectId: 'findy-id',
+  storageBucket: 'findy-id.appspot.com',
+  messagingSenderId: '943881521886'
+}
