+++ src/index.js
@@ -1,11 +1,10 @@
-import firebase from 'firebase/app';
-import 'firebase/auth';
+import firebase from 'firebase/app'
+import 'firebase/auth'
 
 import { firebaseConfig } from './config'
 
 // initialize when loaded
-firebase.initializeApp(firebaseConfig);
+firebase.initializeApp(firebaseConfig)
 
 export * from './FirebaseAuthRepository'
 export { FindyIdComponent } from './FindyIdComponent'
-
